Refactor exec code with stdout capture
