+++ nukeserversocket/controllers/houdini/scripts/python/app/main.py
@@ -1,25 +1,19 @@
 """Nuke-specific plugin for the NukeServerSocket."""
 from __future__ import annotations
 
-import traceback
 from typing import Optional
 
 from PySide2.QtWidgets import QWidget
 
 from nukeserversocket.main import NukeServerSocket
-from nukeserversocket.utils import stdoutIO
+from nukeserversocket.utils import exec_code
 from nukeserversocket.received_data import ReceivedData
 from nukeserversocket.controllers.base import BaseController
 
 
 class HoudiniController(BaseController):
     def execute(self, data: ReceivedData) -> str:
-        with stdoutIO() as s:
-            try:
-                exec(data.text, globals())
-            except Exception:
-                return traceback.format_exc()
-            return s.getvalue()
+        return exec_code(data.text, data.file)
 
 
 class HoudiniEditor(NukeServerSocket):
