+++ nukeserversocket/utils/__init__.py
@@ -1,2 +1,2 @@
 from .cache import cache, clear_cache
-from .stdout import stdoutIO
+from .exec_code import stdoutIO, exec_code
